requirements.txt
tensorflow-macos
tensorflow-metal
opencv-python-headless
pytesseract
Pillow
scipy
matplotlib